Replaces int with size_t

+++ include/ForwardDeclarations.hpp
@@ -1,12 +1,13 @@
 #ifndef FORWARD_DECLARATIONS_H
 #define FORWARD_DECLARATIONS_H
 #include <type_traits>
+#include <cstddef>
 
 namespace StealthTileMap {
     namespace internal {
         template <typename T> struct traits {
             typedef T ScalarType;
-            static constexpr int length = 1, width = 1, height = 1, size = 1;
+            static constexpr size_t length = 1, width = 1, height = 1, size = 1;
         };
     } /* internal */
 
@@ -15,8 +16,8 @@ namespace StealthTileMap {
     class TileMapBase;
 
     // TileMap
-    template <typename type, int widthAtCompileTime, int lengthAtCompileTime = 1, int heightAtCompileTime = 1,
-        int areaAtCompileTime = widthAtCompileTime * lengthAtCompileTime, int sizeAtCompileTime
+    template <typename type, size_t widthAtCompileTime, size_t lengthAtCompileTime = 1, size_t heightAtCompileTime = 1,
+        size_t areaAtCompileTime = widthAtCompileTime * lengthAtCompileTime, size_t sizeAtCompileTime
         = widthAtCompileTime * lengthAtCompileTime * heightAtCompileTime>
     class TileMap;
 
@@ -30,7 +31,7 @@ namespace StealthTileMap {
     class UnaryOp;
 
     // View of a section of a TileMap or OpStruct
-    template <int widthAtCompileTime, int lengthAtCompileTime, int heightAtCompileTime, typename TileMapType,
+    template <size_t widthAtCompileTime, size_t lengthAtCompileTime, size_t heightAtCompileTime, typename TileMapType,
         typename containsData = typename internal::traits<TileMapType>::containsData,
         typename isWritable = typename std::conjunction<containsData, typename internal::traits<TileMapType>::isWritable>::type>
     class TileMapView;
