+++ include/TileMapBase.hpp
@@ -9,39 +9,39 @@ namespace StealthTileMap {
         public:
             typedef typename internal::traits<Derived>::ScalarType ScalarType;
             // Dimensions
-            static constexpr int width() noexcept {
+            static constexpr size_t width() noexcept {
                 return internal::traits<Derived>::width;
             }
 
-            static constexpr int length() noexcept {
+            static constexpr size_t length() noexcept {
                 return internal::traits<Derived>::length;
             }
 
-            static constexpr int height() noexcept {
+            static constexpr size_t height() noexcept {
                 return internal::traits<Derived>::height;
             }
 
-            static constexpr int area() noexcept {
+            static constexpr size_t area() noexcept {
                 return internal::traits<Derived>::area;
             }
 
-            static constexpr int size() noexcept {
+            static constexpr size_t size() noexcept {
                 return internal::traits<Derived>::size;
             }
 
-            constexpr auto operator()(int x, int y, int z) const {
+            constexpr auto operator()(size_t x, size_t y, size_t z) const {
                 return static_cast<const Derived*>(this) -> operator()(x, y, z);
             }
 
-            constexpr auto operator()(int x, int y) const {
+            constexpr auto operator()(size_t x, size_t y) const {
                 return static_cast<const Derived*>(this) -> operator()(x, y);
             }
 
-            constexpr auto operator()(int x) const {
+            constexpr auto operator()(size_t x) const {
                 return static_cast<const Derived*>(this) -> operator()(x);
             }
 
-            constexpr auto operator[](int x) const {
+            constexpr auto operator[](size_t x) const {
                 return static_cast<const Derived*>(this) -> operator[](x);
             }
 
@@ -74,10 +74,10 @@ namespace StealthTileMap {
 
     template <typename Derived>
     std::ostream& operator<<(std::ostream& os, const TileMapBase<Derived>& tileMap) {
-        for (int k = 0; k < tileMap.height(); ++k) {
+        for (size_t k = 0; k < tileMap.height(); ++k) {
             os << "Layer " << k << '\n';
-            for (int j = 0; j < tileMap.length(); ++j) {
-                for (int i = 0; i < tileMap.width(); ++i) {
+            for (size_t j = 0; j < tileMap.length(); ++j) {
+                for (size_t i = 0; i < tileMap.width(); ++i) {
                     os << internal_to_string(tileMap(i, j, k)) << " ";
                 }
                 os << '\n';
